# backend/docs/api.doc.yaml
openapi: 3.1.0
info:
  title: School Hub API
  version: 1.0.0
  description: API documentation for the SchoolHub application
  contact:
    name: SchoolHub Development Team
    email: samitesfaye726@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:5000/api
  description: Development server
paths:
...
  /roles:
    post:
      summary: Create a new role
      description: Creates a new role in the system.
      operationId: createRole
      tags:
      - Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '201':
          description: Role created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleValidationError'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get all roles
      description: Retrieves a list of all roles in the system.
      operationId: getAllRoles
      tags:
      - Roles
      responses:
        '200':
          description: A list of roles.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /roles/{id}:
    get:
      summary: Get a role by ID
      description: Retrieves a specific role by its ID.
      tags:
      - Roles
      operationId: getRoleById
      parameters:
      - name: id
        in: path
        required: true
        description: The ID of the role to retrieve.
        schema:
          type: string
      responses:
        '200':
          description: Role retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '404':
          description: Role not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleNotFoundResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a role
      description: Updates a specific role by its ID.
      tags:
      - Roles
      operationId: updateRole
      parameters:
      - name: id
        in: path
        required: true
        description: The ID of the role to update.
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '200':
          description: Role updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleValidationError'
        '404':
          description: Role not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleNotFoundResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a role
      description: Deletes a specific role by its ID.
      tags:
      - Roles
      operationId: deleteRole
      parameters:
      - name: id
        in: path
        required: true
        description: The ID of the role to delete.
        schema:
        type: string
      responses:
        '204':
          description: Role deleted successfully.
        '404':
          description: Role not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleNotFoundResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Internal server error."
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Validation error."
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "Username is required."
              path:
                type: string
    RoleRequest:
      type: object
      properties:
        name:
          type: string
          example: "Admin"
        description:
          type: string
          example: "Administrator role with full access."
      required:
      - name
    RoleResponse:
      type: object
      properties:
        id:
          type: string
          example: "921126bb-e3b9-47d5-a519-d2fd5348d901"
        name:
          type: string
          example: "Admin"
        description:
          type: string
          example: "Administrator role with full access."
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
...
    RoleNotFoundResponse:
      type: object
      properties:
        message:
          type: string
          example: "Role with ID 9201edf1-a97c-4dc1-8ab2-f3351c7cbffa does not exist."
    RoleValidationError:
      type: object
      properties:
        message:
          type: string
          example: "Validation error."
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "Name is required."
              path:
                type: string
                example: "name"
...